style: move TabContext to Tab component directory
